--- app/src/components/organizations/AuditTrail.vue
@@ -1,187 +1,0 @@
-<template>
-  <v-card class="mb-5">
-    <v-card-title class="d-flex justify-space-between cve-background">
-      Audit Trail
-      <v-switch
-        v-if="trail.length"
-        v-model="viewMode"
-        false-value="table"
-        true-value="timeline"
-        hide-details
-        color="primary"
-        inset
-      >
-        <template v-slot:prepend>
-          <v-icon
-            :color="viewMode === 'table' ? 'primary' : 'grey'"
-            icon="mdi-table"
-          />
-        </template>
-
-        <template v-slot:append>
-          <v-icon
-            :color="viewMode === 'timeline' ? 'primary' : 'grey'"
-            icon="mdi-timeline-outline"
-          />
-        </template>
-      </v-switch>
-    </v-card-title>
-    <v-divider></v-divider>
-    <v-card-text>
-      <div v-if="!hasAudit">
-        <v-btn @click="loadAuditLog()"> Load Audit log </v-btn>
-      </div>
-      <div v-else>
-        <v-timeline side="end" v-if="viewMode === 'timeline'">
-          <v-timeline-item
-            v-for="(s, idx) in visibleTrail"
-            :key="s.id"
-            dot-color="primary"
-          >
-            <template #opposite>
-              <div class="text-caption">{{ s.timestampText }}</div>
-              <div class="text-caption text-medium-emphasis">
-                Author: {{ s.author }}
-              </div>
-            </template>
-            <v-card variant="outlined">
-              <v-card-text v-if="idx === 0"> Creation </v-card-text>
-
-              <v-card-text v-else>
-                <div class="d-flex justify-space-between">
-                  <v-spacer></v-spacer>
-                  <span class="text-disabled">
-                    {{
-                      s.changeCount ? `${s.changeCount} changes` : "No changes"
-                    }}
-                  </span>
-                </div>
-                <v-list density="compact" lines="two">
-                  <v-list-item v-for="c in s.changes" :key="c.path">
-                    <v-list-item-title
-                      ><code>{{ c.path }}</code></v-list-item-title
-                    >
-                    <v-list-item-subtitle>
-                      <span class="text-medium-emphasis">{{
-                        c.beforeText
-                      }}</span>
-                      <span class="mx-2">→</span>
-                      <span>{{ c.afterText }}</span>
-                    </v-list-item-subtitle>
-                  </v-list-item>
-                </v-list>
-
-                <div v-if="!s.changes.length" class="text-medium-emphasis">
-                  No field-level changes detected.
-                </div>
-              </v-card-text>
-            </v-card>
-          </v-timeline-item>
-          <v-timeline-item v-if="hasMore" hide-dot>
-            <v-btn variant="tonal" block @click="loadMore">
-              Load more snapshots
-            </v-btn>
-          </v-timeline-item>
-        </v-timeline>
-
-        <v-data-table
-          :items="trail"
-          :headers="headers"
-          item-value="id"
-          show-expand
-          :expanded="expanded"
-          striped="even"
-          v-if="viewMode === 'table'"
-        >
-          <template #item.changed="{ item }">
-            {{ item.changes.map((c) => c.path).join(", ") }}
-          </template>
-          <template v-slot:expanded-row="{ columns, item }">
-            <td :colspan="columns.length" class="pa-0 ml-5">
-              <v-table density="compact" class="diff-table sub-table">
-                <thead>
-                  <tr>
-                    <!-- <th></th> -->
-                    <th class="bg-grey-lighten-3 text-left w-33">Field</th>
-                    <th class="bg-grey-lighten-3 text-left w-33">Before</th>
-                    <th class="bg-grey-lighten-3 text-left w-33">After</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  <tr v-for="(change, i) in item.changes" :key="i">
-                    <!-- <td></td> -->
-                    <td>
-                      {{ change.path }}
-                    </td>
-                    <td class="text-medium-emphasis">
-                      {{ change.beforeText ?? "—" }}
-                    </td>
-                    <td class="text-success font-weight-medium">
-                      {{ change.afterText ?? "—" }}
-                    </td>
-                  </tr>
-
-                  <tr v-if="!item.changes.length">
-                    <td
-                      colspan="3"
-                      class="text-center text-medium-emphasis py-3"
-                    >
-                      No field-level changes detected.
-                    </td>
-                  </tr>
-                </tbody>
-              </v-table>
-            </td>
-          </template>
-        </v-data-table>
-      </div>
-    </v-card-text>
-  </v-card>
-</template>
-
-<script setup>
-import { useOrgStore } from "@/stores/organizaionStore";
-import { buildAuditTrail } from "./OrgUtils";
-
-import { computed, ref } from "vue";
-
-const PAGE_SIZE = 5;
-const visibleCount = ref(PAGE_SIZE);
-const viewMode = ref("table");
-const store = useOrgStore();
-const expanded = ref([]);
-const hasAudit = computed(() => store.getHasAudit);
-
-const history = computed(() => store.currentOrg?.audit?.history ?? []);
-
-const headers = [
-  { title: "Time", key: "timestampText", sortable: true},
-  { title: "User", key: "author", sortable: false },
-  { title: "Changes", key: "changeCount", sortable: false },
-  { title: "Changed", key: "changed", sortable: false },
-];
-
-const trail = computed(() =>
-  buildAuditTrail(history.value, {
-    ignoredPaths: ["__v", "_id", "last_updated", "lastUpdated"],
-  }).filter((s) => s.changeCount > 0),
-);
-
-const visibleTrail = computed(() => trail.value.slice(0, visibleCount.value));
-
-const hasMore = computed(() => visibleCount.value < trail.value.length);
-
-function loadMore() {
-  visibleCount.value += PAGE_SIZE;
-}
-
-function loadAuditLog() {
-  store.fetchAuditLogsForOrg(store.currentOrg?.org?.short_name);
-}
-</script>
-
-<style scoped>
-.v-timeline-item__body .v-card {
-  width: 350px;
-}
-</style>
