+++ app/src/components/organizations/OrgUtils.js
@@ -1,4 +1,5 @@
 import diff from "microdiff";
+import { useUserStore } from "../../stores/userStore";
 
 export const baseFields = [
   "long_name",
@@ -66,10 +67,24 @@ function formatValue(v) {
   return JSON.stringify(v);
 }
 
+const flattenObject = (obj, prefix = "") => {
+  if (obj == null || typeof obj !== "object" || Array.isArray(obj)) {
+    return [{ path: prefix, value: obj }];
+  }
+  return Object.entries(obj).flatMap(([key, val]) => {
+    const newPath = prefix ? `${prefix}.${key}` : key;
+    if (val != null && typeof val === "object" && !Array.isArray(val)) {
+      return flattenObject(val, newPath);
+    }
+    return [{ path: newPath, value: val }];
+  });
+};
+
 export function buildAuditTrail(
   history,
   { ignoredPaths = ["__v", "_id", "last_updated", "lastUpdated"], locale } = {},
 ) {
+  const userStore = useUserStore();
   const sorted = [...(history ?? [])].sort(
     (a, b) => new Date(a.timestamp) - new Date(b.timestamp),
   );
@@ -101,17 +116,26 @@ export function buildAuditTrail(
 
     const changes =
       prev == null
-        ? []
+        ? flattenObject(curr)
+            .filter(({ path }) => !shouldIgnore(path))
+            .map(({ path, value }) => ({
+              type: "CREATE",
+              path,
+              before: undefined,
+              after: value,
+              beforeText: "—",
+              afterText: formatValue(value),
+            }))
         : diff(prev, curr)
             .map((c) => {
               const pathStr = c.path.join(".");
               const before =
                 c.type === "CREATE" ? undefined : getAtPath(prev, c.path);
               const after =
-                c.type === "REMOVE" ? undefined : getAtPath(curr, c.path); //This may be overkill.
+                c.type === "REMOVE" ? undefined : getAtPath(curr, c.path);
 
               return {
-                type: c.type, // CHANGE | CREATE | REMOVE
+                type: c.type,
                 path: pathStr,
                 before,
                 after,
@@ -121,11 +145,13 @@ export function buildAuditTrail(
             })
             .filter((c) => !shouldIgnore(c.path));
 
+    const _author = userStore.users.find((u) => u.UUID === entry.change_author);
+
     return {
       id: entry._id,
       timestamp: entry.timestamp, // top-level field of the history entry
       timestampText: dtf.format(new Date(entry.timestamp)),
-      author: entry.change_author,
+      author: _author ? _author.username : entry.change_author,
       changes,
       changeCount: changes.length,
     };
