+++ app/src/components/organizations/OrganizationViewPage.vue
@@ -342,7 +342,7 @@ import { computed, onMounted, watch } from "vue";
 import { useRouter } from "vue-router";
 import { useOrgStore } from "../../stores/organizaionStore.js";
 import { useAuthStore } from "../../stores/loginAuth.js";
-import AuditTrail from "./AuditTrail.vue";
+import AuditTrail from "./audits/AuditTrail.vue";
 import Conversations from "./conversations/ConversationView.vue";
 import MembersView from "./MembersView.vue";
 import { OrgUtils } from "./OrgUtils.js";
