+++ app/src/components/organizations/audits/AuditTrail.vue
@@ -1,0 +1,69 @@
+<template>
+  <v-card class="mb-5">
+    <v-card-title class="d-flex justify-space-between cve-background">
+      Audit Trail
+      <v-switch
+        v-if="trail.length"
+        v-model="viewMode"
+        false-value="table"
+        true-value="timeline"
+        hide-details
+        color="primary"
+        inset
+      >
+        <template v-slot:prepend>
+          <v-icon
+            :color="viewMode === 'table' ? 'primary' : 'grey'"
+            icon="mdi-table"
+          />
+        </template>
+
+        <template v-slot:append>
+          <v-icon
+            :color="viewMode === 'timeline' ? 'primary' : 'grey'"
+            icon="mdi-timeline-outline"
+          />
+        </template>
+      </v-switch>
+    </v-card-title>
+    <v-divider></v-divider>
+    <v-card-text>
+      <div v-if="!hasAudit" class="d-flex justify-center">
+        <EmptyField></EmptyField>
+      </div>
+      <div v-else>
+        <TableStyle :trail="trail" v-if="viewMode === 'table'" />
+        <TimelineStyle :trail="trail" v-else />
+      </div>
+    </v-card-text>
+  </v-card>
+</template>
+
+<script setup>
+import { useOrgStore } from "@/stores/organizaionStore";
+import { buildAuditTrail } from "../OrgUtils";
+import { computed, onBeforeMount, ref } from "vue";
+import TimelineStyle from "./TimelineStyle.vue";
+import TableStyle from "./TableStyle.vue";
+import { useUserStore } from "../../../stores/userStore";
+
+const viewMode = ref("table");
+const store = useOrgStore();
+const userStore = useUserStore();
+
+const hasAudit = computed(() => store.getHasAudit);
+
+const history = computed(() => store.currentOrg?.audit?.history ?? []);
+
+const trail = computed(() =>
+  buildAuditTrail(history.value, {
+    ignoredPaths: ["__v", "_id", "last_updated", "lastUpdated"],
+  }).filter((s) => s.changeCount > 0),
+);
+
+onBeforeMount(async () => {
+  if (userStore.users.length === 0) {
+    await userStore.fetchUsers();
+  }
+});
+</script>
