+++ app/src/components/organizations/audits/TableStyle.vue
@@ -1,0 +1,68 @@
+<template>
+  <v-data-table
+    :items="trail"
+    :headers="headers"
+    item-value="id"
+    show-expand
+    :expanded="expanded"
+    striped="even"
+  >
+    <template #item.changed="{ item }">
+      {{ item.changes.map((c) => c.path).join(", ") }}
+    </template>
+    <template v-slot:expanded-row="{ columns, item }">
+      <td :colspan="columns.length" class="pa-0 ml-5">
+        <v-table density="compact" class="diff-table sub-table">
+          <thead>
+            <tr>
+              <!-- <th></th> -->
+              <th class="bg-grey-lighten-3 text-left w-33">Field</th>
+              <th class="bg-grey-lighten-3 text-left w-33">Before</th>
+              <th class="bg-grey-lighten-3 text-left w-33">After</th>
+            </tr>
+          </thead>
+          <tbody>
+            <tr v-for="(change, i) in item.changes" :key="i">
+              <!-- <td></td> -->
+              <td>
+                {{ change.path }}
+              </td>
+              <td class="text-medium-emphasis">
+                {{ change.beforeText ?? "—" }}
+              </td>
+              <td class="text-success font-weight-medium">
+                {{ change.afterText ?? "—" }}
+              </td>
+            </tr>
+
+            <tr v-if="!item.changes.length">
+              <td colspan="3" class="text-center text-medium-emphasis py-3">
+                No field-level changes detected.
+              </td>
+            </tr>
+          </tbody>
+        </v-table>
+      </td>
+    </template>
+  </v-data-table>
+</template>
+
+<script setup>
+import { computed, ref } from "vue";
+
+const headers = [
+  { title: "Time", key: "timestampText", sortable: true },
+  { title: "User", key: "author", sortable: false },
+  { title: "Changes", key: "changeCount", sortable: false },
+  { title: "Changed", key: "changed", sortable: false },
+];
+
+const expanded = ref([]);
+
+const props = defineProps({
+  trail: {
+    type: Object,
+    required: true,
+  },
+});
+</script>
