+++ app/src/components/organizations/audits/TimelineStyle.vue
@@ -1,0 +1,73 @@
+<template>
+  <v-timeline side="end">
+    <v-timeline-item
+      v-for="(s, idx) in visibleTrail"
+      :key="s.id"
+      dot-color="primary"
+    >
+      <template #opposite>
+        <div class="text-caption">{{ s.timestampText }}</div>
+        <div class="text-caption text-medium-emphasis">
+          Author: {{ s.author }}
+        </div>
+      </template>
+      <v-card variant="outlined">
+        <v-card-text>
+          <div class="d-flex justify-space-between">
+            <v-spacer></v-spacer>
+            <span class="text-disabled">
+              {{ s.changeCount ? `${s.changeCount} changes` : "No changes" }}
+            </span>
+          </div>
+          <v-list density="compact" lines="two">
+            <v-list-item v-for="c in s.changes" :key="c.path">
+              <v-list-item-title
+                ><code>{{ c.path }}</code></v-list-item-title
+              >
+              <v-list-item-subtitle>
+                <span class="text-medium-emphasis">{{ c.beforeText }}</span>
+                <span class="mx-2">→</span>
+                <span>{{ c.afterText }}</span>
+              </v-list-item-subtitle>
+            </v-list-item>
+          </v-list>
+
+          <div v-if="!s.changes.length" class="text-medium-emphasis">
+            No field-level changes detected.
+          </div>
+        </v-card-text>
+      </v-card>
+    </v-timeline-item>
+    <v-timeline-item v-if="hasMore" hide-dot>
+      <v-btn variant="tonal" block @click="loadMore">
+        Load more snapshots
+      </v-btn>
+    </v-timeline-item>
+  </v-timeline>
+</template>
+
+<script setup>
+import { computed, ref } from "vue";
+
+const props = defineProps({
+  trail: {
+    type: Object,
+    required: true,
+  },
+});
+
+const PAGE_SIZE = 5;
+const visibleCount = ref(PAGE_SIZE);
+
+const visibleTrail = computed(() => props.trail.slice(0, visibleCount.value));
+const hasMore = computed(() => visibleCount.value < props.trail.length);
+
+function loadMore() {
+  visibleCount.value += PAGE_SIZE;
+}
+</script>
+<style scoped>
+.v-timeline-item__body .v-card {
+  width: 350px;
+}
+</style>
