+++ app/src/stores/organizaionStore.js
@@ -96,13 +96,16 @@ export const useOrgStore = defineStore("org", () => {
 
   async function fetchOrgUserByUsername(orgShortName, username) {
     try {
-      const user = await authStore.orgApi.fetchUserForOrg(orgShortName, username);
+      const user = await authStore.orgApi.fetchUserForOrg(
+        orgShortName,
+        username,
+      );
       // Hydrate with org info
-      const orgAndRole = getOrgAndRoleForUser(user.UUID)
+      const orgAndRole = getOrgAndRoleForUser(user.UUID);
       return {
         ...user,
-        ...orgAndRole
-      }
+        ...orgAndRole,
+      };
     } catch (e) {
       console.log(e);
     }
@@ -121,6 +124,7 @@ export const useOrgStore = defineStore("org", () => {
       // keep same behavior (fire-and-forget)
       fetchOrgUsersByIdentifier(identifier);
       fetchReviewForOrg(identifier);
+      fetchAuditLogsForOrg(identifier);
 
       return currentOrg;
     } catch (e) {
